Add files via upload
alea is an encryption program. It works on the basis of a transposition on the bit level Transposition is the mixing up of the parts of a message.
Take for examplethe name "Fiedel Castro" you could change the order of the name "Friedel Tosac" from it
just by changing the sequence of letters. Likewise, you can only use individual components of a text